+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Binance parser: added cash deposit and withdrawal exports.
 - Binance parser: added statements export.
 - Bitfinex parser: new "Trades" data file format added. ([#41](https://github.com/BittyTax/BittyTax/issues/41))
+- Bittrex parser: new deposits data file format added.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
 - Binance parser: use filename to determine if deposits or withdrawals.